restructure classes by package
