+++ warcard/src/com/svi/main/PlayerCard2.java
