+++ warcard/src/com/svi/process/ConvertSuitAndRank.java
@@ -1,4 +1,4 @@
-package com.svi.warcard;
+package com.svi.process;
 import java.util.StringTokenizer;
 
 public class ConvertSuitAndRank {
