+++ warcard/src/com/svi/process/ProcessGame.java
@@ -1,9 +1,12 @@
-package com.svi.warcard;
+package com.svi.process;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.LinkedHashSet;
 import java.util.StringTokenizer;
+
+import com.svi.object.MsterInfo;
+import com.svi.object.RoundInfo;
 
 
 public class ProcessGame {
