+++ warcard/src/com/svi/process/ReadInputTextFile.java
@@ -1,4 +1,4 @@
-package com.svi.warcard;
+package com.svi.process;
 import java.io.*;
 import java.util.ArrayList;
 import java.util.StringTokenizer;
