+++ warcard/src/com/svi/process/ShuffleCards.java
@@ -1,4 +1,4 @@
-package com.svi.warcard;
+package com.svi.process;
 import java.util.ArrayList;
 
 public class ShuffleCards {
